COM-2013 - change every function in activities selectores to cards selectors
