+++ src/screens/courses/cardTemplates/MultipleChoiceQuestionCard/index.tsx
@@ -5,7 +5,7 @@ import shuffle from 'lodash/shuffle';
 import { useNavigation } from '@react-navigation/native';
 import { footerColorsType, MultipleChoiceQuestionType, qcmAnswerFromAPIType } from '../../../../types/CardType';
 import { StateType } from '../../../../types/store/StoreType';
-import Selectors from '../../../../store/activities/selectors';
+import Selectors from '../../../../store/cards/selectors';
 import Actions from '../../../../store/activities/actions';
 import CardHeader from '../../../../components/cards/CardHeader';
 import { GREEN, GREY, ORANGE, PINK } from '../../../../styles/colors';
