+++ src/store/cards/selectors.ts
@@ -1,4 +1,4 @@
-import { OPEN_QUESTION, QUESTION_ANSWER, SURVEY } from '../../core/data/constants';
+import { OPEN_QUESTION, QUESTION_ANSWER, SURVEY, TRANSITION } from '../../core/data/constants';
 import { QuestionnaireAnswerType } from '../../types/store/CardStoreType';
 
 const getCard = state => state.cards.cards[state.cards.cardIndex];
@@ -9,7 +9,21 @@ const getQuestionnaireAnswer = (state): QuestionnaireAnswerType | null => {
   return state.cards.questionnaireAnswersList.find(qa => qa.card === card._id) || null;
 };
 
+const getMaxProgress = state => state.cards.cards.filter(card => card.template !== TRANSITION).length;
+
+const getProgress = (state) => {
+  const { cards, cardIndex } = state.cards;
+  if (!Number.isInteger(cardIndex)) return 0;
+
+  return 1 + cards.filter(c => c.template !== TRANSITION).map(c => c._id).indexOf(cards[cardIndex]._id);
+};
+
+const displayProgressBar = state => !!getCard(state) && getCard(state).template !== TRANSITION;
+
 export default {
   getCard,
   getQuestionnaireAnswer,
+  getMaxProgress,
+  getProgress,
+  displayProgressBar,
 };
